medical_app_django/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap5==2024.2
Django==4.2.13
django-crispy-forms==2.1
exceptiongroup==1.2.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
mysqlclient==2.2.4
packaging==24.0
pip==24.0
playwright==1.44.0
pluggy==1.5.0
pyee==11.1.0
pytest==8.2.1
requests==2.32.2
setuptools==69.5.1
soupsieve==2.5
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0